Fetch commit trees, split github module
